+++ src/qilisdk/speqtrum/speqtrum.py
@@ -24,11 +24,9 @@ import httpx
 from loguru import logger
 from pydantic import TypeAdapter
 
-from qilisdk.functionals.sampling import Sampling
-from qilisdk.functionals.time_evolution import TimeEvolution
-from qilisdk.functionals.variational_program import VariationalProgram
+from qilisdk.functionals import Sampling, TimeEvolution, VariationalProgram
 from qilisdk.settings import get_settings
-from qilisdk.speqtrum.experiments.experiment_functional import ExperimentFunctional, RabiExperiment, T1Experiment
+from qilisdk.speqtrum.experiments import ExperimentFunctional, RabiExperiment, T1Experiment
 
 from .keyring import delete_credentials, load_credentials, store_credentials
 from .speqtrum_models import (
